==> analysis/03_Gene_Inspection.Rmd <==
---
title: "Check Markers"
author: "AGC, MJ"
date: "27 08 2021"
output: html_document
---

```{r}

home = getwd()
```

```{r setup, include=FALSE}

output= paste0(home, "/output/")
rankconz=T
source(paste0(home,"/code/custom_functions.R"))

library(DESeq2)
library(knitr)
library(tidyverse)
library(pheatmap)

opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dpi=500
)



```

```{r load_data}

load(paste0(output,"/dds_matrix.RData"))

SampleInfo = as.data.frame(colData(ddsMat))
genedata=rowData(ddsMat)

SampleInfo = SampleInfo %>% dplyr::arrange(Drug, Conz, CellLine, Batch)

```


```{r plotGOIS, fig.height=15, fig.width=10}

GOIsEntrezFull=c(NCAM1="4684",
                 EGF="1950",
                 EGFR="1956",
                 FGFR1="2260",
                 FGFR2="2263",
                 FGFR3="2261",
                 FGFR4="2264",
                 FGF1="2246",
                 FGF2="2247",
                 TRKC="4916",
                 NT3="4908",
                 TRKB="4915",
                 NGFR="4804",
                 PRKACA="5566",
                 PRKACB="5567",
                 PRKAR1A="5573",
                 PRKAR2B="5577",
                 PRKCA="5578",
                 OCT4="5460",
                 NANOG="79923",
                 MSI1="4440",
                 SOX1="6656",
                 TBR1=	"10716",
                 TBR2=	"8320",
                 MASH1_ASCL1="429",
                 HES5="388585",
                 PAX6="5080",
                 SOX2="6657",
                 NESTIN="10763",
                 GFAP="2670",
                 S100B="6285",
                 AQP4="361",
                 SOX9="6662",
                 CX3CR1="1524",
                 OLIG2="10215",
                 GRIA2="2891",
                 GRIN1	="2902",
                 GRM1="2911",
                 GRIN2A="2903",
                 VGLUT1="57030",
                 GLAST_SLC1A3="6507",
                 GLUL="2752",
                 GAD1="2571",	
                 GAT1=	"6529",	
                 ENO2=	"2026",	
                 MAPT=	"4137",	
                 DCX="1641",
                 MAP2="4133",
                 TUBB3="10381",
                 NEUN="146713",
                 PSD95="1742",
                 HOMER1="9456",
                 ACHE=	"43",
                 HPRT1=	"3251",
                 NET1=	"10276",
                 POU3F2=	"5454",
                 TH="7054",
                 SYNAPTOPHYSIN="6855",
                 GAP_43="2596",
                 BASSOON="8927",
                 STXBP1="6812",
                 STX1A="6804",
                 SNCA="6622")

GOIsEntrez=c(
  OCT4="5460",
  NANOG="79923",
  SOX2="6657",
  NESTIN="10763",
  PAX6="5080",
  GFAP="2670",
  AQP4="361",
  ASCL1="429",
  Olig2="10215",
  TMEM119="338773",
  TUBB3="10381",
  MAP2="4133",
  SYP="6855",
  SLC17A7="57030",
  GAD1="2571",
  SERT1="6532",
  CHAT="1103",
  TH="7054")


geneheatmap=function(GOIsEntrez,exprobj,CellID, rankconz=F){
  
  idx = match(GOIsEntrez, rownames(exprobj))
  vsd <- vst(exprobj, blind=FALSE)
  mat <- assay(vsd)
  log_2cpm <- limma::removeBatchEffect(mat, vsd$Batch, vsd$CellLine)

  idxsmpl = SampleInfo$CellLine %in% CellID
  
  dataset= log_2cpm[idx, rownames(SampleInfo)[idxsmpl]]
  
  colnames(dataset)= SampleInfo[rownames(SampleInfo)[idxsmpl],"label_rep"]
  rownames(dataset) = names(GOIsEntrez)
  
  #colors for plotting heatmap
  colors <- rev(colorRampPalette(brewer.pal(9, "Spectral"))(255))
  
  cellcol = Dark8[1:nlevels(SampleInfo$CellLine)]
  names(cellcol) = levels(SampleInfo$CellLine)
  
  drugcol = Dark8[c(1:nlevels(SampleInfo$Drug)+nlevels(SampleInfo$CellLine))]
  names(drugcol) = levels(SampleInfo$Drug)
  
  
  batchcol = Dark8[c(1:nlevels(SampleInfo$Batch)+nlevels(SampleInfo$CellLine)+nlevels(SampleInfo$Drug))]
  names(batchcol) = levels(SampleInfo$Batch)
  
  conzcol = c("orange", "darkred")
  
  ann_colors = list(
    CellLine = cellcol, 
    Drug = drugcol,
    Batch = batchcol,
    Conz=conzcol)
  
  
  labels = SampleInfo[,c("Drug","Conz","CellLine","Batch")] %>%  
    mutate_if(is.factor, as.character) %>% as.data.frame()
   
  if(rankconz){
    labels$Conz = as.numeric(factor(labels$Conz, levels=c(unique(labels$Conz)), 
                          labels=c(0,1,2,1,2,1,2)))
   }
  
  rownames(labels)= colnames(dataset)
  
  pheatmap(dataset, 
           cluster_rows = F,
           cluster_cols = F,
           col = colors,
           scale = "none", 
           annotation_col = labels,
           annotation_colors = ann_colors, 
           main=CellID)
}

geneheatmap(GOIsEntrez = GOIsEntrez,
            exprobj = ddsMat,
            CellID=unique(SampleInfo$CellLine),
            rankconz = T)


```
