==> analysis/04_WGCNA.Rmd <==
---
title: "WGCNA"
author: "AGC, AY"
date: "22 3 2021"
output: html_document
---

```{r}
home = getwd()
```


```{r setup, include=FALSE}
reanalyze=F
rankconz=T

output= paste0(home, "/output/")

source(paste0(home,"/code/custom_functions.R"))

library(DESeq2)
library(limma)
library(knitr)
library(tidyverse)
library(pheatmap)
library(WGCNA)
library(flashClust)
library(pheatmap)
library(gprofiler2)
library(xlsx)
library(lme4)


opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dpi=500
)


load(paste0(output,"/dds_matrix.RData"))
SampleInfo = as.data.frame(colData(ddsMat))

```

# WGCNA

## Identify Gene Modules

```{r adjacency}

#calclute the adjacency matrix
allowWGCNAThreads()

vsd <- vst(ddsMat, blind=FALSE)
vsd <- assay(vsd)
vsd <- limma::removeBatchEffect(vsd, ddsMat$Batch, ddsMat$CellLine)

WGCNA_matrix <- t(vsd) #Need to transform for further calculations


if(reanalyze | !file.exists(paste0(output,"/WGCNA_adj_TOM.RData"))){
  
  s = abs(bicor(WGCNA_matrix)) #biweight mid-correlation
  powers = c(c(1:10), seq(from = 12, to=20, by=2))
  sft = pickSoftThreshold(WGCNA_matrix, powerVector = powers, verbose = 5)
  
  par(mfrow = c(1,2))
  cex1 = 0.9;
  plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
       xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit, signed R^2",
       type="n", main = paste("Scale independence"));
  text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
       labels=powers,cex=cex1,col="red");
  abline(h=0.90,col="red")
  
  plot(sft$fitIndices[,1], sft$fitIndices[,5],xlab="Soft Threshold (power)",ylab="Mean Connectivity", 
       type="n",main = paste("Mean connectivity"))
  text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
  
  softPower = 8; # The point where the curve flattens
  
  adj= adjacency(WGCNA_matrix,type = "unsigned", power = softPower);
  
  #Converting adjacency matrix into To so that the noise could be reduced
  TOM=TOMsimilarityFromExpr(WGCNA_matrix,networkType = "unsigned", 
                            TOMType = "unsigned", power = softPower);
  save(list = c("adj", "TOM"), file=paste0(output,"/WGCNA_adj_TOM.RData"))
  
} else {
  print("loading precalculated data")
  load(paste0(output,"/WGCNA_adj_TOM.RData"))
}

SubGeneNames<-colnames(WGCNA_matrix)

colnames(TOM) =rownames(TOM) =SubGeneNames
dissTOM=1-TOM
diag(dissTOM) = 0

#hierarchical clustering
geneTree = flashClust(as.dist(dissTOM),method="average");

#plot the resulting clustering tree (dendrogram)
# plot(geneTree, xlab="", sub="",cex=0.3, main="Module clustering prio to merging");

#Set the minimum module size
minModuleSize = 50;

#Module identification using dynamic tree cut
dynamicMods = cutreeDynamic(dendro = geneTree, 
                            distM = dissTOM, 
                            cutHeight = 0.998,
                            minClusterSize = minModuleSize,
                            deepSplit=2, 
                            pamRespectsDendro = FALSE)

#the following command gives the module labels and the size of each module. Lable 0 is reserved for unassigned genes

#Plot the module assignment under the dendrogram; note: The grey color is reserved for unassigned genes
dynamicColors = labels2colors(dynamicMods)
table(dynamicColors)

plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut", 
                    dendroLabels = FALSE, hang = 0.03, 
                    addGuide = TRUE, 
                    guideHang = 0.05, 
                    main = "Gene dendrogram and module colors")

diag(dissTOM) = NA;

#Visualize the Tom plot. Raise the dissimilarity matrix to the power of 4 to bring out the module structure
TOMplot(dissTOM^4, geneTree, as.character(dynamicColors), main="weighted distance of Topological overlap Matrix")

```

# Module eigengenes 
## Identify and Merge correlated modules 

```{r MEs,  fig.height=20, fig.width=15}
#calculate eigengenes
MEList = moduleEigengenes(WGCNA_matrix, colors = dynamicColors)
MEs = MEList$eigengenes 
plotEigengeneNetworks(MEs, "Eigengenes_before_merging", marDendro = c(0,4,1,2), marHeatmap = c(3,4,1,2))

MEList_new = mergeCloseModules(WGCNA_matrix, 
                               colors = dynamicColors, 
                               MEs = MEs, 
                               cutHeight = 0.1)

plotEigengeneNetworks(MEList_new$newMEs, "Eigengenes_after_merging", 
                      marDendro = c(0,4,1,2), marHeatmap = c(3,4,1,2))

MEs = MEList_new$newMEs

colData(ddsMat) = colData(ddsMat) %>% cbind(MEs[colnames(ddsMat),])

colors_new = MEList_new$colors

mcols(ddsMat) = cbind(mcols(ddsMat), data.frame(cluster=colors_new))

clusterdataframe=data.frame(entrezgene=mcols(ddsMat)$entrez_gene,hgnc=mcols(ddsMat)$hgnc,  cluster=colors_new)

save(ddsMat, file=paste0(output,"/dds_matrix.RData"))
write.csv(clusterdataframe, file=paste0(output,"/clusterdataframe.csv"))

```

## Heatmaps MEs
```{r heatmap,  fig.height=20, fig.width=15}
#colors for plotting heatmap

colors <- rev(colorRampPalette(brewer.pal(9, "Spectral"))(255))

cellcol = Dark8[1:nlevels(SampleInfo$CellLine)]
names(cellcol) = levels(SampleInfo$CellLine)

drugcol = Dark8[c(1:nlevels(SampleInfo$Drug)+nlevels(SampleInfo$CellLine))]
names(drugcol) = levels(SampleInfo$Drug)

batchcol = Dark8[c(1:nlevels(SampleInfo$Batch)+nlevels(SampleInfo$CellLine)+nlevels(SampleInfo$Drug))]
names(batchcol) = levels(SampleInfo$Batch)

conzcol = c("orange", "darkred")

clustercol = unique(colors_new)
names(clustercol) = clustercol

ann_colors = list(
  CellLine = cellcol, 
  Drug = drugcol,
  Batch = batchcol,
  Conz=conzcol, 
  cluster = clustercol)

idx=order(SampleInfo$CellLine, SampleInfo$Drug, SampleInfo$Conz,SampleInfo$Batch)
WGCNA_matrix_sorted=WGCNA_matrix[idx, order(colors_new)]
rownames(WGCNA_matrix_sorted) = SampleInfo$label_rep[idx]

collabels = SampleInfo[idx,c("CellLine","Drug","Conz","Batch")] %>%  
  mutate_if(is.factor, as.character) %>% as.data.frame()
if(rankconz){
    collabels$Conz = as.numeric(factor(collabels$Conz, levels=c(unique(collabels$Conz)), 
                          labels=c(0,1,2,1,2,1,2)))
   }
rownames(collabels)=SampleInfo$label_rep[idx]

genlabels = data.frame(cluster = as.character(colors_new)[order(colors_new)])
rownames(genlabels) = colnames(WGCNA_matrix_sorted)

pheatmap(t(WGCNA_matrix_sorted),
         cluster_rows= F,
         cluster_cols = T,
         border_color = NA, 
         clustering_method = "ward.D2",
         annotation_row = genlabels,
         annotation_col = collabels,
         annotation_colors = ann_colors,
         col = colors, 
         scale = "row",
         main = "Distances  normalized log2 counts")

MElabels = data.frame(cluster = gsub("ME", "",colnames(MEs)))
rownames(MElabels) = colnames(MEs)

rownames(MEs) = SampleInfo[rownames(MEs),"label_rep"]

pheatmap(t(MEs[idx,]),
         border_color = NA, 
         annotation_row = MElabels,
         cluster_cols = T,
         annotation_col = collabels,
         clustering_method = "ward.D2",
         annotation_colors = ann_colors,
         scale="none",
         col = colors, 
         main = "Distances  normalized log2 counts")


```


```{r goresplots}
gene_univers = rownames(ddsMat)

Genes_of_interset = split(rownames(ddsMat), mcols(ddsMat)$cluster)

gostres = getGOresults(Genes_of_interset, gene_univers)

toptab = gostres$result

write.xlsx2(toptab, file = paste0(output,"/GOresWGCNA.xlsx"), sheetName = "GO_enrichment")

for (module in names(Genes_of_interset)){
  idx = toptab$query==module & grepl("GO", toptab$source)
  if(!any(idx)){
    p = ggplot() + annotate("text", x = 4, y = 25, size=4, 
                            label = "no significant GO term") + 
      ggtitle(module)+theme_void()+
      theme(plot.title = element_text(hjust = 0.5))
  } else {
    p=GOplot(toptab[idx, ], 10, Title =module)
  }
  print(p)
}


```


```{r stats_ME}

SampleInfo=as.data.frame(colData(ddsMat))
MEnames = grep("ME", colnames(SampleInfo), value = T)
randomeffect = c("Batch", "CellLine")

target = "Conz"

for (drug in unique(SampleInfo$Drug)){
  res=data.frame()
  for (mod in MEnames){
    re = paste0("(1|", randomeffect, ")", collapse="+")
    y = SampleInfo[,mod]
    model = paste0("y~",re,"+", target)
    data = SampleInfo[SampleInfo$Drug==drug,c(randomeffect, target)]
    y=y[SampleInfo$Drug==drug]
    resultlme = lmer(model, data)
    stats = summary(resultlme)$coeff[2,]
    res = rbind(res, stats)
    colnames(res) = colnames(summary(resultlme)$coeff)
  }
  rownames(res) = MEnames
  res$p.value = 2*pt(res$`t value`, sum(SampleInfo$Drug==drug), lower=FALSE)
  res$p.value[res$p.value>1]=1
  assign(paste0("WGCNA_lmer_", drug), res)
}

# 
# 
# 
# boxplot(SampleInfo[SampleInfo$Drug=="Levetiracetam", "MEcyan"] ~ as.factor(SampleInfo[SampleInfo$Drug=="Valporate", "Conz"]), 
#          ylab="MEcyan", xlab="mM", col="cyan")

for (i in apropos("WGCNA_lmer_")){
  print(i)
  tmp=get(i)
  print(tmp)
  write.csv(tmp, file=paste0(output, "/",i,".csv"))
  xlsx::write.xlsx(tmp, file=paste0(output, "/",i,".xlsx"))
  if( any(tmp$p.value <=0.05)){
    Drug=gsub("WGCNA_lmer_","",i)
    for(mod in rownames(tmp)[tmp$p.value<=0.1]){
    boxplot(SampleInfo[SampleInfo$Drug==Drug, mod] ~
              as.factor(SampleInfo[SampleInfo$Drug==Drug, "Conz"]), 
            ylab=mod, xlab=paste("mM", Drug), col=gsub("ME","",mod))
      mtext(paste("p.value =",tmp[mod,"p.value"]), 3)
    }
    
  }
  
}

save(list=apropos("WGCNA_lmer_"), file=paste0(output,"/WGCNA_lmer_res.RData"))


```
