==> analysis/05_Statistical_analysis.Rmd <==
---
title: "Statistical analysis"
author: "AGC, AY"
date: "23 3 2021"
output: html_document
---

# Statistical analysis 

```{r setup, include=FALSE, echo=F}
reanalyze=T

home = getwd()
output= paste0(home, "/output/")

source(paste0(home,"/code/custom_functions.R"))

library(DESeq2)
library(knitr)
library(tidyverse)
library(pheatmap)
library(WGCNA)
library(flashClust)
library(pheatmap)
library(limma)
library(xlsx)

opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dpi=500
)


load(paste0(output,"/dds_matrix.RData"))
SampleInfo = as.data.frame(colData(ddsMat))

vsd <- vst(ddsMat, blind=FALSE)
vsd <- assay(vsd)
vsd <- limma::removeBatchEffect(vsd, ddsMat$Batch, ddsMat$CellLine)
log_2cpm=vsd


```



# Q1 Which genes are differentially regualted upon Drug Concentration 


```{r loops_for_comparison, message=F}

if(reanalyze | !file.exists(paste0(home,"/output/ResTabs_DE.RData"))){
  # calculate all combinations 
  
  Drugs=unique(as.character(SampleInfo$Drug))
  target="Conz"
  # no random effects included 
  for(d in Drugs){
    Drugfilter = SampleInfo$Drug %in% d
    Set = rownames(SampleInfo)[Drugfilter]
    lab = paste("restab", d, sep="_")
    print(lab)
    assign(lab, 
           comparison(ddsMat, samples = Set, 
                      target =target,randomeffect = c("Batch", "CellLine")))
  }
  comparisons= apropos("restab")
  save(list = comparisons, file = paste0(home,"/output/ResTabs_KO.RData"))
  
} else 
  load(file = paste0(home,"/output/ResTabs_DE.RData"))


```


```{r predef_for_analaysis}
mypval=0.05
nompval = 0.001

colors <- rev(colorRampPalette(brewer.pal(9, "Spectral"))(255))

getlinmodoutput = function(targetdrug){
  
  samplesincl = SampleInfo$Drug==targetdrug 
  res = get(paste0("restab_", targetdrug))
  pvalrep =  res$adj.P.Val<mypval 
  hits=rownames(res)[pvalrep]
  nomp=F
  corrtest="adjp"
  
  if(length(hits)==0){
    nomp=T
    pvalrep =  res$P.Value<nompval
    hits=rownames(res)[pvalrep]
    print("nominal P-value used")
    corrtest="uncorr_sig"
  }
  
  SamplesSet=SampleInfo[samplesincl,]
  
  plotmatrix = log_2cpm[hits,rownames(SamplesSet)]
  
  genlist=mcols(ddsMat)[hits,c("hgnc", "entrez_gene")]
  write.csv(genlist, paste0(output,"/",targetdrug,"_",corrtest,".csv"))
  
  rownames(SamplesSet)=SamplesSet$label_rep
  
  colnames(plotmatrix)=SamplesSet$label_rep
  rownames(plotmatrix)=rowData(ddsMat)[hits,"hgnc"]     
  
  cellcol = Dark8[1:nlevels(SampleInfo$CellLine)]
  names(cellcol) = levels(SampleInfo$CellLine)
  
  drugcol = Dark8[c(1:nlevels(SampleInfo$Drug)+nlevels(SampleInfo$CellLine))]
  names(drugcol) = levels(SampleInfo$Drug)
  
  
  batchcol = Dark8[c(1:nlevels(SampleInfo$Batch)+nlevels(SampleInfo$CellLine)+nlevels(SampleInfo$Drug))]
  names(batchcol) = levels(SampleInfo$Batch)
  
  conzcol = c("orange", "darkred")
  
  ann_colors = list(
    CellLine = cellcol, 
    Drug = drugcol,
    Batch = batchcol,
    Conz=conzcol)
  
  collabels = SamplesSet[,c("CellLine","Drug","Conz","Batch")] %>%  
    mutate_if(is.factor, as.character) %>% as.data.frame()
  
  rownames(collabels)=SamplesSet$label_rep
  maintitle = "Normalized log2 counts PROLIF noRAPA"
  if(nomp){maintitle=paste(maintitle, "nominal sig")}
  
  pheatmap(plotmatrix[,],
           border_color = NA, 
           annotation_col = collabels[,],
           annotation_colors = ann_colors,
           clustering_method = "ward.D2",
           cluster_cols = F,
           col = colors, 
           scale = "row",
           main = maintitle)
  
  
  gene_univers = rownames(ddsMat)
  
  gostres = getGOresults(hits, gene_univers)
  
  toptab = gostres$result
  
  
  write.xlsx2(toptab, file = paste0(output, "/GOres_",targetdrug,".xlsx"), sheetName = "GO_enrichment")
  
  idx = grepl("GO|KEGG", toptab$source)
  titlename = paste(targetdrug)
  if(!any(idx)){
    p = ggplot() + annotate("text", x = 4, y = 25, size=4, 
                            label = "no significant GO term") + 
      ggtitle(titlename)+theme_void()+
      theme(plot.title = element_text(hjust = 0.5))
  } else {
    p=GOplot(toptab[idx, ], 10, Title = titlename)
  }
  print(p)
}


```

## Drug Effects
### Carbamazepine

```{r Carbamazepine,  fig.height=20, fig.width=15}
getlinmodoutput(targetdrug = "Carbamazepine")

```


### Lacosamide 

```{r Lacosamide ,  fig.height=20, fig.width=15}

getlinmodoutput(targetdrug = "Lacosamide")

```

### Levetiracetam  

```{r Levetiracetam  ,  fig.height=20, fig.width=15}

getlinmodoutput(targetdrug = "Levetiracetam")

```

### Valporate 

```{r Valporate ,  fig.height=20, fig.width=15}

getlinmodoutput(targetdrug = "Valproate")

```
